+++ src/configui/app/util/eufy-security-client.utils.ts
@@ -68,6 +68,23 @@ export class Device {
     return false;
   }
 
+  static isBatteryDoorbell(type: number): boolean {
+    if (type === DeviceType.BATTERY_DOORBELL ||
+        type === DeviceType.BATTERY_DOORBELL_2 ||
+        type === DeviceType.BATTERY_DOORBELL_PLUS
+    ) {
+      return true;
+    }
+    return false;
+  }
+
+  static isWiredDoorbell(type: number): boolean {
+    if (type === DeviceType.DOORBELL) {
+      return true;
+    }
+    return false;
+  }
+
   static isCamera(type: number): boolean {
     if (
       type === DeviceType.CAMERA ||