+++ src/plugin/accessories/DoorbellCameraAccessory.ts
@@ -1,4 +1,4 @@
-import { Service, PlatformAccessory, DoorbellOptions } from 'homebridge';
+import { Service, PlatformAccessory, DoorbellOptions, CharacteristicValue } from 'homebridge';
 
 import { EufySecurityPlatform } from '../platform';
 
@@ -6,7 +6,7 @@ import { EufySecurityPlatform } from '../platform';
 
 // eslint-disable-next-line @typescript-eslint/ban-ts-comment
 // @ts-ignore  
-import { DoorbellCamera, Device } from 'eufy-security-client';
+import { DoorbellCamera, Device, PropertyName } from 'eufy-security-client';
 import { CameraAccessory } from './CameraAccessory';
 
 /**
@@ -20,6 +20,8 @@ export class DoorbellCameraAccessory extends CameraAccessory {
   private ring_triggered: boolean;
 
   private doorbellService: Service;
+
+  private indoorChimeSwitchService?: Service;
 
   constructor(
     platform: EufySecurityPlatform,
@@ -64,6 +66,39 @@ export class DoorbellCameraAccessory extends CameraAccessory {
     }
 
     this.doorbellService.setPrimaryService(true);
+
+    // add indoor chime switch
+    try {
+      if ((this.eufyDevice.isBatteryDoorbell() || this.eufyDevice.isWiredDoorbell()) && this.cameraConfig.indoorChimeButton) {
+        this.platform.log.debug(this.accessory.displayName, 'indoorChimeSwitch config:', this.cameraConfig.indoorChimeButton);
+        this.platform.log.debug(this.accessory.displayName, 'has a indoorChime, so append indoorChimeSwitchService to it.');
+
+        this.indoorChimeSwitchService =
+          this.accessory.getService('indoorChimeSwitch') || 
+          this.accessory.addService(this.platform.Service.Switch, 'indoorChimeSwitch', 'indoorChime');
+        
+        this.indoorChimeSwitchService.setCharacteristic(this.platform.Characteristic.Name, this.accessory.displayName + ' indoor chime');
+        // this.indoorChimeSwitchService.setCharacteristic(this.platform.Characteristic.ConfiguredName,
+        //   this.accessory.displayName + ' indoor chime');
+
+        this.indoorChimeSwitchService.getCharacteristic(this.characteristic.On)
+          .onGet(this.handleIndoorChimeGet.bind(this))
+          .onSet(this.handleIndoorChimeSet.bind(this));
+      } else {
+        this.platform.log.debug(this.accessory.displayName,
+          'Looks like not compatible with indoorChime or this has been disabled within configuration');
+
+        // remove indoorChimeButton service if the user has disabled the it through the config
+        this.indoorChimeSwitchService = accessory.getService('indoorChimeSwitch');
+        if (this.indoorChimeSwitchService) {
+          this.platform.log.debug(this.accessory.displayName, 'removing indoorChimeSwitch service.');
+          accessory.removeService(this.indoorChimeSwitchService);
+        }
+      }
+
+    } catch (err) {
+      this.platform.log.error(this.accessory.displayName, 'raise error in indoorChimeSwitchService.', err);
+    }
   }
 
   // We receive 2 push when Doorbell ring, mute the second by checking if we already send
@@ -80,4 +115,24 @@ export class DoorbellCameraAccessory extends CameraAccessory {
     }
   }
 
+  async handleIndoorChimeGet(): Promise<CharacteristicValue> {
+    try {
+      const currentValue = this.eufyDevice.getPropertyValue(PropertyName.DeviceChimeIndoor);
+      this.platform.log.debug(this.accessory.displayName, 'GET DeviceChimeIndoor:', currentValue);
+      return currentValue as boolean;
+    } catch {
+      this.platform.log.debug(this.accessory.displayName, 'handleIndoorChimeGet', 'Wrong return value');
+      return false;
+    }
+  }
+
+  async handleIndoorChimeSet(value: CharacteristicValue) {
+    try {
+      this.platform.log.debug(this.accessory.displayName, 'SET DeviceChimeIndoor:', value);
+      const station = await this.platform.getStationById(this.eufyDevice.getStationSerial());
+      await station.enableIndoorChime(this.eufyDevice, value as boolean);
+    } catch (err) {
+      this.platform.log.debug(this.accessory.displayName, 'handleIndoorChimeSet error', err);
+    }
+  }
 }