+++ src/plugin/accessories/StationAccessory.ts
@@ -98,8 +98,13 @@ export class StationAccessory {
     
     this.manualTriggerService.setCharacteristic(
       this.characteristic.Name,
-      accessory.displayName + ' alarm switch',
+      accessory.displayName + ' alarm',
     );
+
+    // this.manualTriggerService.setCharacteristic(
+    //   this.platform.Characteristic.ConfiguredName,
+    //   accessory.displayName + ' alarm',
+    // );
 
     this.manualTriggerService
       .getCharacteristic(this.characteristic.On)