+++ src/plugin/utils/configTypes.ts
@@ -23,9 +23,11 @@ export type CameraConfig = {
   immediateRingNotificationWithoutSnapshot?: boolean;
   delayCameraSnapshot?:boolean;
   talkback?: boolean;
+  talkbackChannels?: number;
   hsv?: boolean;
   hsvRecordingDuration?: number;
   experimentalRTSP?: boolean;
+  indoorChimeButton?: boolean;
 };
 
 export type VideoConfig = {